[DREALODEMA-123] give responsibility to format columns and datas to TableTabulator component
